+++ tests/CMapTest.cpp
@@ -26,26 +26,26 @@ using namespace std;
 
 static void check_range (const SegmentedCMap &cmap, int min, int max, int cid_min) {
 	for (int i=min; i <= max; i++)
-		ASSERT_EQ(cmap.cid(i), cid_min+(i-min));
+		ASSERT_EQ(cmap.cid(i), uint32_t(cid_min+(i-min)));
 }
 
 
 static void check_zero (const SegmentedCMap &cmap, int min, int max) {
 	for (int i=min; i <= max; i++)
-		ASSERT_EQ(cmap.cid(i), 0);
+		ASSERT_EQ(cmap.cid(i), 0u);
 }
 
 
 TEST(SegmentedCMapTest, disjoint_ranges) {
 	SegmentedCMap cmap("test");
 	cmap.addCIDRange(5, 8, 1);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 5, 8, 1);
 	check_zero(cmap, 0, 4);
 	check_zero(cmap, 9, 20);
 
 	cmap.addCIDRange(10, 15, 5);
-	ASSERT_EQ(cmap.numCIDRanges(), 2);
+	ASSERT_EQ(cmap.numCIDRanges(), 2u);
 	check_range(cmap, 5, 8, 1);
 	check_range(cmap, 10, 15, 5);
 	check_zero(cmap, 0, 4);
@@ -53,7 +53,7 @@ TEST(SegmentedCMapTest, disjoint_ranges) {
 	check_zero(cmap, 16, 20);
 
 	cmap.addCIDRange(0, 3, 50);
-	ASSERT_EQ(cmap.numCIDRanges(), 3);
+	ASSERT_EQ(cmap.numCIDRanges(), 3u);
 	check_range(cmap, 0, 3, 50);
 	check_range(cmap, 5, 8, 1);
 	check_range(cmap, 10, 15, 5);
@@ -62,7 +62,7 @@ TEST(SegmentedCMapTest, disjoint_ranges) {
 	check_zero(cmap, 16, 20);
 
 	cmap.addCIDRange(16, 20, 1);
-	ASSERT_EQ(cmap.numCIDRanges(), 4);
+	ASSERT_EQ(cmap.numCIDRanges(), 4u);
 	check_range(cmap, 5, 8, 1);
 	check_range(cmap, 10, 15, 5);
 	check_range(cmap, 0, 3, 50);
@@ -73,20 +73,20 @@ TEST(SegmentedCMapTest, disjoint_ranges) {
 TEST(SegmentedCMapTest, touching_ranges) {
 	SegmentedCMap cmap("test");
 	cmap.addCIDRange(5, 8, 10);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 5, 8, 10);
 
 	cmap.addCIDRange(9, 15, 14);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 5, 15, 10);
 
 	cmap.addCIDRange(1, 4, 5);
-	ASSERT_EQ(cmap.numCIDRanges(), 2);
+	ASSERT_EQ(cmap.numCIDRanges(), 2u);
 	check_range(cmap, 1, 4, 5);
 	check_range(cmap, 5, 15, 10);
 
 	cmap.addCIDRange(1, 4, 6);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 1, 15, 6);
 }
 
@@ -95,26 +95,26 @@ TEST(SegmentedCMapTest, overlapping_ranges) {
 	SegmentedCMap cmap("test");
 	cmap.addCIDRange(5, 8, 10);
 	cmap.addCIDRange(7, 15, 12);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 5, 15, 10);
 
 	cmap.addCIDRange(10, 20, 1);
-	ASSERT_EQ(cmap.numCIDRanges(), 2);
+	ASSERT_EQ(cmap.numCIDRanges(), 2u);
 	check_range(cmap, 5, 9, 10);
 	check_range(cmap, 10, 20, 1);
 
 	cmap.addCIDRange(2, 7, 7);
-	ASSERT_EQ(cmap.numCIDRanges(), 2);
+	ASSERT_EQ(cmap.numCIDRanges(), 2u);
 	check_range(cmap, 2, 9, 7);
 	check_range(cmap, 10, 20, 1);
 
 	cmap.addCIDRange(1, 12, 100);
-	ASSERT_EQ(cmap.numCIDRanges(), 2);
+	ASSERT_EQ(cmap.numCIDRanges(), 2u);
 	check_range(cmap, 1, 12, 100);
 	check_range(cmap, 13, 20, 4);
 
 	cmap.addCIDRange(0, 30, 1);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 0, 30, 1);
 }
 
@@ -123,24 +123,24 @@ TEST(SegmentedCMapTest, inner_ranges) {
 	SegmentedCMap cmap("test");
 	cmap.addCIDRange(5, 20, 1);
 	cmap.addCIDRange(10, 15, 6);
-	ASSERT_EQ(cmap.numCIDRanges(), 1);
+	ASSERT_EQ(cmap.numCIDRanges(), 1u);
 	check_range(cmap, 5, 20, 1);
 
 	cmap.addCIDRange(10, 15, 100);
-	ASSERT_EQ(cmap.numCIDRanges(), 3);
+	ASSERT_EQ(cmap.numCIDRanges(), 3u);
 	check_range(cmap, 5, 9, 1);
 	check_range(cmap, 10, 15, 100);
 	check_range(cmap, 16, 20, 12);
 
 	cmap.addCIDRange(15, 15, 50);
-	ASSERT_EQ(cmap.numCIDRanges(), 4);
+	ASSERT_EQ(cmap.numCIDRanges(), 4u);
 	check_range(cmap, 5, 9, 1);
 	check_range(cmap, 10, 14, 100);
 	check_range(cmap, 15, 15, 50);
 	check_range(cmap, 16, 20, 12);
 
 	cmap.addCIDRange(6, 19, 1);
-	ASSERT_EQ(cmap.numCIDRanges(), 3);
+	ASSERT_EQ(cmap.numCIDRanges(), 3u);
 	check_range(cmap, 5, 5, 1);
 	check_range(cmap, 6, 19, 1);
 	check_range(cmap, 20, 20, 16);
