+++ tests/RangeMapTest.cpp
@@ -26,26 +26,26 @@ using namespace std;
 
 static void check_range (const RangeMap &rangemap, int min, int max, int minval) {
 	for (int i=min; i <= max; i++)
-		ASSERT_EQ(rangemap.valueAt(i), minval+(i-min));
+		ASSERT_EQ((int)rangemap.valueAt(i), minval+(i-min));
 }
 
 
 static void check_zero (const RangeMap &rangemap, int min, int max) {
 	for (int i=min; i <= max; i++)
-		ASSERT_EQ(rangemap.valueAt(i), 0);
+		ASSERT_EQ(rangemap.valueAt(i), 0u);
 }
 
 
 TEST(RangeMapTest, disjoint_ranges) {
 	RangeMap rangemap;
 	rangemap.addRange(5, 8, 1);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 5, 8, 1);
 	check_zero(rangemap, 0, 4);
 	check_zero(rangemap, 9, 20);
 
 	rangemap.addRange(10, 15, 5);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 5, 8, 1);
 	check_range(rangemap, 10, 15, 5);
 	check_zero(rangemap, 0, 4);
@@ -53,7 +53,7 @@ TEST(RangeMapTest, disjoint_ranges) {
 	check_zero(rangemap, 16, 20);
 
 	rangemap.addRange(0, 3, 50);
-	ASSERT_EQ(rangemap.size(), 3);
+	ASSERT_EQ(rangemap.size(), 3u);
 	check_range(rangemap, 0, 3, 50);
 	check_range(rangemap, 5, 8, 1);
 	check_range(rangemap, 10, 15, 5);
@@ -62,7 +62,7 @@ TEST(RangeMapTest, disjoint_ranges) {
 	check_zero(rangemap, 16, 20);
 
 	rangemap.addRange(16, 20, 1);
-	ASSERT_EQ(rangemap.size(), 4);
+	ASSERT_EQ(rangemap.size(), 4u);
 	check_range(rangemap, 5, 8, 1);
 	check_range(rangemap, 10, 15, 5);
 	check_range(rangemap, 0, 3, 50);
@@ -73,20 +73,20 @@ TEST(RangeMapTest, disjoint_ranges) {
 TEST(RangeMapTest, touching_ranges1) {
 	RangeMap rangemap;
 	rangemap.addRange(5, 8, 10);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 5, 8, 10);
 
 	rangemap.addRange(9, 15, 14);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 5, 15, 10);
 
 	rangemap.addRange(1, 4, 5);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 1, 4, 5);
 	check_range(rangemap, 5, 15, 10);
 
 	rangemap.addRange(1, 4, 6);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 1, 15, 6);
 }
 
@@ -94,16 +94,16 @@ TEST(RangeMapTest, touching_ranges1) {
 TEST(RangeMapTest, touching_ranges2) {
 	RangeMap rangemap;
 	rangemap.addRange(7, 8, 10);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 7, 8, 10);
 
 	rangemap.addRange(2, 3, 1);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 2, 3, 1);
 	check_range(rangemap, 7, 8, 10);
 
 	rangemap.addRange(4, 4, 3);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 2, 4, 1);
 	check_range(rangemap, 7, 8, 10);
 }
@@ -113,26 +113,26 @@ TEST(RangeMapTest, overlapping_ranges) {
 	RangeMap rangemap;
 	rangemap.addRange(5, 8, 10);
 	rangemap.addRange(7, 15, 12);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 5, 15, 10);
 
 	rangemap.addRange(10, 20, 1);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 5, 9, 10);
 	check_range(rangemap, 10, 20, 1);
 
 	rangemap.addRange(2, 7, 7);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 2, 9, 7);
 	check_range(rangemap, 10, 20, 1);
 
 	rangemap.addRange(1, 12, 100);
-	ASSERT_EQ(rangemap.size(), 2);
+	ASSERT_EQ(rangemap.size(), 2u);
 	check_range(rangemap, 1, 12, 100);
 	check_range(rangemap, 13, 20, 4);
 
 	rangemap.addRange(0, 30, 1);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 0, 30, 1);
 }
 
@@ -141,24 +141,24 @@ TEST(RangeMapTest, inner_ranges) {
 	RangeMap rangemap;
 	rangemap.addRange(5, 20, 1);
 	rangemap.addRange(10, 15, 6);
-	ASSERT_EQ(rangemap.size(), 1);
+	ASSERT_EQ(rangemap.size(), 1u);
 	check_range(rangemap, 5, 20, 1);
 
 	rangemap.addRange(10, 15, 100);
-	ASSERT_EQ(rangemap.size(), 3);
+	ASSERT_EQ(rangemap.size(), 3u);
 	check_range(rangemap, 5, 9, 1);
 	check_range(rangemap, 10, 15, 100);
 	check_range(rangemap, 16, 20, 12);
 
 	rangemap.addRange(15, 15, 50);
-	ASSERT_EQ(rangemap.size(), 4);
+	ASSERT_EQ(rangemap.size(), 4u);
 	check_range(rangemap, 5, 9, 1);
 	check_range(rangemap, 10, 14, 100);
 	check_range(rangemap, 15, 15, 50);
 	check_range(rangemap, 16, 20, 12);
 
 	rangemap.addRange(6, 19, 1);
-	ASSERT_EQ(rangemap.size(), 3);
+	ASSERT_EQ(rangemap.size(), 3u);
 	check_range(rangemap, 5, 5, 1);
 	check_range(rangemap, 6, 19, 1);
 	check_range(rangemap, 20, 20, 16);
